Add gru4rec package modules
