+++ gru4rec/__init__.py
@@ -1,0 +1,13 @@
+"""GRU4Rec package exposing core classes."""
+from .data import SessionDataIterator
+from .model import GRUEmbedding, GRU4RecModel
+from .optimizers import IndexedAdagradM
+from .trainer import GRU4Rec
+
+__all__ = [
+    "IndexedAdagradM",
+    "GRUEmbedding",
+    "GRU4RecModel",
+    "SessionDataIterator",
+    "GRU4Rec",
+]
